+++ backend/canisters/group/api/src/updates/c2c_join_group.rs
@@ -1,6 +1,6 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
-use types::{GateCheckFailedReason, GroupCanisterGroupChatSummary, TimestampMillis, UserId};
+use types::{GateCheckFailedReason, GroupCanisterGroupChatSummary, TimestampMillis, UserId, VerifiedCredentialGateArgs};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -11,6 +11,7 @@ pub struct Args {
     pub is_platform_moderator: bool,
     pub is_bot: bool,
     pub diamond_membership_expires_at: Option<TimestampMillis>,
+    pub verified_credential_args: Option<VerifiedCredentialGateArgs>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]